Car control improvement
Extrapolating on the use of my velocity decay in creating spring stiffness while grounded I replaced the rigid bodies drag with a decay of the z axis 2% of speed is lost per physics tick while accelerating and 5% is lost while not accelerating. The 2% causes the car to have a max speed that it can achieve and realistically this value should not be messed with. the 5% can be see as the stopping speed with 5% being a rather quick stop. This might be changed to create slower stopping vehicles.

Made minor changes to space key mechanic